Ajout: Page login et Map intéractive

+++ polytech/src/main/java/com/univtours/polytech/entity/Reservation.java
@@ -14,44 +14,44 @@ public class Reservation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reservation")
-    private Integer reservation;
+    private Long reservation;
 
     @Column(name = "utilisateur_id")
-    private Integer utilisateur_id;
+    private Long utilisateur_id;
 
     @Column(name = "terrain_id")
-    private Integer terrain_id;
+    private Long terrain_id;
 
     public Reservation() {
     }
 
-    public Reservation(Integer reservation, Integer utilisateur_id, Integer terrain_id) {
+    public Reservation(Long reservation, Long utilisateur_id, Long terrain_id) {
         this.reservation = reservation;
         this.utilisateur_id = utilisateur_id;
         this.terrain_id = terrain_id;
     }
 
-    public Integer getReservation() {
+    public Long getReservation() {
         return reservation;
     }
 
-    public void setReservation(Integer reservation) {
+    public void setReservation(Long reservation) {
         this.reservation = reservation;
     }
 
-    public Integer getUtilisateur_id() {
+    public Long getUtilisateur_id() {
         return utilisateur_id;
     }
 
-    public void setUtilisateur_id(Integer utilisateur_id) {
+    public void setUtilisateur_id(Long utilisateur_id) {
         this.utilisateur_id = utilisateur_id;
     }
 
-    public Integer getTerrain_id() {
+    public Long getTerrain_id() {
         return terrain_id;
     }
 
-    public void setTerrain_id(Integer terrain_id) {
+    public void setTerrain_id(Long terrain_id) {
         this.terrain_id = terrain_id;
     }
 }
